Separate the parser from the caching. Massive rewrite of the Builder system to avoid redundant code (hence 3.0)
