+++ CHANGELOG
@@ -156,6 +156,8 @@ v2.2
 - Detect Firefox Focus and a more generic pattern for WebView applications
 - Fixed NPE
 
-v2.3
+v3.0
 ===
 - Avoid DDOS (too long parsing times). Without actual parsing: Useragent too tong == hacker.
+- Separate the parser from the caching.
+- Massive rewrite of the Builder system to avoid redundant code (hence 3.0)
