+++ ReleaseProcedure.txt
@@ -8,8 +8,8 @@ mvn clean verify -Pplatfora
 mvn clean
 
 # Prepare the versions for the release
-RELEASEVERSION=2.2
-NEXTVERSION=2.3
+RELEASEVERSION=3.0
+NEXTVERSION=3.1
 
 # Creating and publishing the new Tagging
 mvn versions:set -DgenerateBackupPoms=false -DnewVersion=${RELEASEVERSION}
