+++ analyzer/src/test/java/nl/basjes/parse/useragent/parse/TestUserAgentFlattening.java
@@ -19,7 +19,7 @@ package nl.basjes.parse.useragent.parse;
 
 import nl.basjes.parse.useragent.UserAgent;
 import nl.basjes.parse.useragent.UserAgentAnalyzer;
-import nl.basjes.parse.useragent.UserAgentAnalyzer.GetAllPathsAnalyzer;
+import nl.basjes.parse.useragent.UserAgentAnalyzerDirect.GetAllPathsAnalyzer;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
