+++ commandline/pom.xml
@@ -24,7 +24,7 @@
   <parent>
     <artifactId>yauaa-parent</artifactId>
     <groupId>nl.basjes.parse.useragent</groupId>
-    <version>2.3-SNAPSHOT</version>
+    <version>3.0-SNAPSHOT</version>
   </parent>
   <artifactId>yauaa-commandline</artifactId>
