+++ devtools/pom.xml
@@ -22,7 +22,7 @@
   <parent>
     <artifactId>yauaa-parent</artifactId>
     <groupId>nl.basjes.parse.useragent</groupId>
-    <version>2.3-SNAPSHOT</version>
+    <version>3.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
