+++ pom.xml
@@ -25,7 +25,7 @@
   <artifactId>yauaa-parent</artifactId>
   <packaging>pom</packaging>
   <groupId>nl.basjes.parse.useragent</groupId>
-  <version>2.3-SNAPSHOT</version>
+  <version>3.0-SNAPSHOT</version>
 
   <name>Yauaa : </name>
   <description>A parsing and analyzing library to get information from a useragent string.</description>
